split rig build tests into module
