+++ pipeline/pipe/m/rig_builder/test/__init__.py
@@ -1,0 +1,10 @@
+from . import core
+from .core import RigBuildTest, TestRunner
+from .tests import RIG_BUILD_TESTS
+
+__all__ = [
+    "core",
+    "RIG_BUILD_TESTS",
+    "RigBuildTest",
+    "TestRunner",
+]
